[MISC] Deprecate views::take_until
- `seqan3::views::take_until` is moved to `seqan3/io/detail` for
  internal use only. As an alternative please use
  `std::views::take_while` in combination with `std::not_fn`.
- moved snippets and test accordingly.
- add CHANGLELOG entry
