+++ include/seqan3/io/structure_file/format_vienna.hpp
@@ -32,13 +32,13 @@
 #include <seqan3/io/detail/istreambuf_view.hpp>
 #include <seqan3/io/detail/misc.hpp>
 #include <seqan3/io/detail/take_line_view.hpp>
+#include <seqan3/io/detail/take_until_view.hpp>
 #include <seqan3/io/detail/take_view.hpp>
 #include <seqan3/io/structure_file/detail.hpp>
 #include <seqan3/io/structure_file/input_format_concept.hpp>
 #include <seqan3/io/structure_file/input_options.hpp>
 #include <seqan3/io/structure_file/output_format_concept.hpp>
 #include <seqan3/io/structure_file/output_options.hpp>
-#include <seqan3/range/views/take_until.hpp>
 #include <seqan3/utility/char_operations/predicate.hpp>
 #include <seqan3/utility/detail/type_name_as_string.hpp>
 #include <seqan3/utility/views/to.hpp>
@@ -140,7 +140,7 @@ protected:
                 if (options.truncate_ids)
                 {
                     std::ranges::copy(stream_view | std::views::drop_while(is_id || is_blank) // skip leading >
-                                                  | views::take_until_or_throw(is_cntrl || is_blank)
+                                                  | detail::take_until_or_throw(is_cntrl || is_blank)
                                                   | views::char_to<std::ranges::range_value_t<id_type>>,
                                       std::cpp20::back_inserter(id));
                     detail::consume(stream_view | detail::take_line_or_throw);
@@ -229,7 +229,7 @@ protected:
         }
         else
         {
-            detail::consume(stream_view | views::take_until(is_space)); // until whitespace
+            detail::consume(stream_view | detail::take_until(is_space)); // until whitespace
         }
 
         if constexpr (!detail::decays_to_ignore_v<seq_type> &&
@@ -259,7 +259,7 @@ protected:
         {
             detail::consume(stream_view | detail::take_line);
         }
-        detail::consume(stream_view | views::take_until(!is_space));
+        detail::consume(stream_view | detail::take_until(!is_space));
     }
 
     //!\copydoc seqan3::structure_file_output_format::write_structure_record
@@ -366,7 +366,7 @@ private:
     auto read_structure(stream_view_type & stream_view)
     {
         auto constexpr is_legal_structure = char_is_valid_for<alph_type>;
-        return stream_view | views::take_until(is_space) // until whitespace
+        return stream_view | detail::take_until(is_space) // until whitespace
                            | std::views::transform([is_legal_structure](char const c)
                              {
                                  if (!is_legal_structure(c))
