+++ test/performance/range/views/CMakeLists.txt
@@ -2,4 +2,3 @@ seqan3_benchmark(view_all_benchmark.cpp)
 seqan3_benchmark(view_drop_benchmark.cpp)
 seqan3_benchmark(view_drop_view_take_benchmark.cpp)
 seqan3_benchmark(view_take_benchmark.cpp)
-seqan3_benchmark(view_take_until_benchmark.cpp)
