+++ test/unit/io/detail/CMakeLists.txt
@@ -7,4 +7,5 @@ seqan3_test(ignore_output_iterator_test.cpp)
 seqan3_test(record_like_test.cpp)
 seqan3_test(safe_filesystem_entry_test.cpp)
 seqan3_test(take_line_view_test.cpp)
+seqan3_test(take_until_view_test.cpp)
 seqan3_test(take_view_test.cpp)
