+++ test/unit/search/views/minimiser_test.cpp
@@ -13,8 +13,8 @@
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/alphabet/views/complement.hpp>
 #include <seqan3/core/detail/debug_stream_alphabet.hpp>
+#include <seqan3/io/detail/take_until_view.hpp>
 #include <seqan3/range/views/drop.hpp>
-#include <seqan3/range/views/take_until.hpp>
 #include <seqan3/search/views/kmer_hash.hpp>
 #include <seqan3/search/views/minimiser.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
@@ -211,7 +211,7 @@ TEST_F(minimiser_test, window_too_big)
 
 TEST_F(minimiser_test, combinability)
 {
-    auto stop_at_t = seqan3::views::take_until([] (seqan3::dna4 const x) { return x == 'T'_dna4; });
+    auto stop_at_t = seqan3::detail::take_until([] (seqan3::dna4 const x) { return x == 'T'_dna4; });
     EXPECT_RANGE_EQ(result3_ungapped_stop, text3 | stop_at_t | kmer_view | minimiser_no_rev_view);
     EXPECT_RANGE_EQ(result3_gapped_stop, text3 | stop_at_t | gapped_kmer_view | minimiser_no_rev_view);
